Put TopicRule in a separate module
